apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic>=2.9.2
pydantic-settings
supabase==2.28.0
cryptography==43.0.1
reportlab==4.2.2
apscheduler==3.10.4
pyjwt>=2.9.0
langgraph==0.2.60
langchain-groq==0.2.1
langchain-core==0.3.28
httpx>=0.24,<0.28
Pillow>=10.0.0
python-multipart==0.0.22
